+++ tasty-plutus/src/Test/Tasty/Plutus/Internal/Minting.hs
@@ -1,23 +1,32 @@
 module Test.Tasty.Plutus.Internal.Minting (
   -- * Types
-  MintingPolicyQuery (..),
+  MintingPolicyTask (..),
+  MintingPolicyAction (..),
+  Tokens (..),
 
   -- * Helpers
-  processMintingQuery,
-
-  -- * Functions
-  burningTokens,
-  mintingTokens,
+  burnTokens,
+  mintTokens,
 ) where
 
 import Data.List.NonEmpty (NonEmpty ((:|)))
 import Plutus.V1.Ledger.Api (TokenName)
-import PlutusTx.Positive (Positive, getPositive)
+import PlutusTx.Positive (Positive)
+import Test.QuickCheck (
+  Arbitrary (arbitrary, shrink),
+  CoArbitrary (coarbitrary),
+  Function (function),
+  functionMap,
+ )
+import Test.QuickCheck.Plutus.Instances ()
 import Prelude
 
-data MintingPolicyQuery
-  = MintQuery TokenName Positive
-  | BurnQuery TokenName Positive
+{- | The task for the tested minting policy.
+
+ @since 6.0
+-}
+data MintingPolicyTask
+  = MPTask MintingPolicyAction Tokens
   deriving stock
     ( -- | @since 6.0
       Eq
@@ -25,21 +34,51 @@ data MintingPolicyQuery
       Show
     )
 
-processMintingQuery :: MintingPolicyQuery -> (TokenName, Integer)
-processMintingQuery = \case
-  MintQuery tn pos -> (tn,) . getPositive $ pos
-  BurnQuery tn pos -> (tn,) . negate . getPositive $ pos
+{- | 'MintingPolicyAction' defines a list of actions for the minting policy.
 
-{- | Create `MintingPolicyQuery` for minting tokens
+ @since 6.0
+-}
+data MintingPolicyAction
+  = BurnAction
+  | MintAction
+  deriving stock (Eq, Show)
+
+{- | 'Tokens' corresponds to some positive number of 'TokenName' belonging
+ to the tested minting policy.
+
+ @since 6.0
+-}
+data Tokens = Tokens TokenName Positive
+  deriving stock (Eq, Show)
+
+-- | @since 6.0
+instance Arbitrary Tokens where
+  arbitrary = Tokens <$> arbitrary <*> arbitrary
+  shrink (Tokens tn pos) = uncurry Tokens <$> shrink (tn, pos)
+
+-- | @since 6.0
+instance CoArbitrary Tokens where
+  coarbitrary (Tokens tn pos) = coarbitrary (tn, pos)
+
+-- | @since 6.0
+instance Function Tokens where
+  function = functionMap into from
+    where
+      into :: Tokens -> (TokenName, Positive)
+      into (Tokens tn pos) = (tn, pos)
+      from :: (TokenName, Positive) -> Tokens
+      from = uncurry Tokens
+
+{- | Create single `MPTask` for minting tokens
 
   @since 6.0
 -}
-mintingTokens :: TokenName -> Positive -> NonEmpty MintingPolicyQuery
-mintingTokens tn i = MintQuery tn i :| []
+mintTokens :: Tokens -> NonEmpty MintingPolicyTask
+mintTokens toks = MPTask MintAction toks :| []
 
-{- | Create `MintingPolicyQuery` for burning tokens
+{- | Create single `MPTask` for burning tokens
 
   @since 6.0
 -}
-burningTokens :: TokenName -> Positive -> NonEmpty MintingPolicyQuery
-burningTokens tn i = BurnQuery tn i :| []
+burnTokens :: Tokens -> NonEmpty MintingPolicyTask
+burnTokens toks = MPTask BurnAction toks :| []