+++ tasty-plutus/src/Test/Tasty/Plutus/TestData.hs
@@ -11,13 +11,15 @@
 module Test.Tasty.Plutus.TestData (
   -- * Data type
   TestData (..),
+  Tokens (..),
   Outcome (..),
-  MintingPolicyQuery (..),
+  MintingPolicyAction (..),
+  MintingPolicyTask (..),
 
   -- * Helper functions
   passIf,
-  burningTokens,
-  mintingTokens,
+  burnTokens,
+  mintTokens,
 
   -- * QuickCheck support
   Methodology (..),
@@ -39,9 +41,11 @@ import Test.Tasty.Plutus.Internal.Context (
   Purpose (ForMinting, ForSpending),
  )
 import Test.Tasty.Plutus.Internal.Minting (
-  MintingPolicyQuery (BurnQuery, MintQuery),
-  burningTokens,
-  mintingTokens,
+  MintingPolicyAction (BurnAction, MintAction),
+  MintingPolicyTask (MPTask),
+  Tokens (Tokens),
+  burnTokens,
+  mintTokens,
  )
 import Prelude
 
@@ -81,7 +85,7 @@ data TestData (p :: Purpose) where
     -- | List of tokens to be minted by the script.
     --
     -- @since 4.1
-    NonEmpty MintingPolicyQuery ->
+    NonEmpty MintingPolicyTask ->
     TestData ( 'ForMinting redeemer)
 
 {- | Describes whether a case, comprised of a script and a test data for the
@@ -216,15 +220,15 @@ data TestItems (p :: Purpose) where
     ) =>
     { -- | Redeemer provided to the MintingPolicy
       -- @since 5.0
-      mintRedeemer :: redeemer
-    , -- | Tokens minted with the MintingPolicy
-      -- @since 5.0
-      mintTokens :: NonEmpty MintingPolicyQuery
+      mpRedeemer :: redeemer
+    , -- | List of MintingPolicy tasks to mint/burn tokens
+      -- @since 6.0
+      mpTasks :: NonEmpty MintingPolicyTask
     , -- | ContextBuilder used for creating ScriptContext
       -- @since 5.0
-      mintCB :: ContextBuilder ( 'ForMinting redeemer)
+      mpCB :: ContextBuilder ( 'ForMinting redeemer)
     , -- | Result expected from calling the MintingPolicy
       -- | @since 5.0
-      mintOutcome :: Outcome
+      mpOutcome :: Outcome
     } ->
     TestItems ( 'ForMinting redeemer)